requirements.txt
asgiref==3.2.10
astroid==2.5.2
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
Django==3.1.2
django-pwa==1.0.10
django-restframework==0.0.1
djangorestframework==3.12.2
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
package-name==0.1
Pillow==8.0.1
pylint==2.7.4
pytz==2020.1
requests==2.25.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
